Move Euler-angle kinematics into euler_angle.rs
The Euler-angle rate maps and the Diebel sequence and rate-matrix
helpers they alone use now sit with the EulerAngle representation, the
same way the quaternion kinematics sit with Quaternion. Their tests move
with them, the attitude module leaves the public paths unchanged, and
the emptied kinematics module is removed.
