+++ src/attitude/mod.rs
@@ -8,7 +8,6 @@ pub mod attitude_types;
 pub(crate) mod common;
 pub mod euler_angle;
 pub mod euler_axis;
-pub mod kinematics;
 pub mod quaternion;
 pub mod rotation_matrix;
 pub mod traits;
@@ -16,7 +15,6 @@ pub mod traits;
 pub use attitude_types::*;
 pub use euler_angle::*;
 pub use euler_axis::*;
-pub use kinematics::*;
 pub use quaternion::*;
 pub use rotation_matrix::*;
 pub use traits::*;
